+++ plugins/voxelman/client/servercontrol.d
@@ -1,0 +1,20 @@
+/**
+Copyright: Copyright (c) 2016 Andrey Penechko.
+License: $(WEB boost.org/LICENSE_1_0.txt, Boost License 1.0).
+Authors: Andrey Penechko.
+*/
+module voxelman.client.servercontrol;
+
+import core.atomic;
+
+shared bool serverShouldStop;
+
+bool isServerRunning()
+{
+	return !atomicLoad(serverShouldStop);
+}
+
+void stopServer()
+{
+	atomicStore(serverShouldStop, true);
+}